created new file for election to reduce organizer
